requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==44.0.3
distro==1.9.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.0
frozendict==2.4.6
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
jsonpatch==1.33
jsonpointer==2.1
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycosat==0.6.6
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==1.7.1
pyparsing==3.2.3
PySocks==1.7.1
python-decouple==3.8
requests==2.32.3
rich==14.0.0
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
setuptools==75.8.0
sqlparse==0.5.3
tqdm==4.67.1
truststore==0.10.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
wheel==0.45.1
zstandard==0.23.0
